👷 Clean up operations in hub after failed runs

+++ tests/storage/conftest.py
@@ -1,4 +1,5 @@
 import shutil
+from pathlib import Path
 from subprocess import DEVNULL, run
 from time import perf_counter
 
@@ -33,21 +34,29 @@ def pytest_sessionstart():
     t_execute_start = perf_counter()
 
     ln_setup._TESTING = True
-    pgurl = setup_local_test_postgres()
+    try:
+        pgurl = setup_local_test_postgres()
+    except RuntimeError:
+        run("docker stop pgtest && docker rm pgtest", shell=True, stdout=DEVNULL)  # noqa: S602
+        pgurl = setup_local_test_postgres()
     try:
         create_test_instance(pgurl)
     except Exception as e:
         print("failed to create test instance:", e)
+        print("deleting the instance")
         delete_test_instance()
-        print("retrying after deleting the instance")
-        create_test_instance(pgurl)
+        # below currently fails because cannot create two instances in the same session
+        # create_test_instance(pgurl)
+        print("now rerun")
+        quit()
     total_time_elapsed = perf_counter() - t_execute_start
     print(f"time to setup the instance: {total_time_elapsed:.1f}s")
 
 
 def delete_test_instance():
     logger.set_verbosity(1)
-    shutil.rmtree("./default_storage_unit_storage")
+    if Path("./default_storage_unit_storage").exists():
+        shutil.rmtree("./default_storage_unit_storage")
     # handle below better in the future
     if ln.UPath("s3://lamindb-test/storage/.lamindb").exists():
         ln.UPath("s3://lamindb-test/storage/.lamindb").rmdir()
@@ -55,12 +64,12 @@ def delete_test_instance():
     if another_storage.exists():
         another_storage.rmdir()
     ln.setup.delete("lamindb-unit-tests-storage", force=True)
-    run("docker stop pgtest && docker rm pgtest", shell=True, stdout=DEVNULL)  # noqa: S602
-    ln.setup.settings.auto_connect = AUTO_CONNECT
 
 
 def pytest_sessionfinish(session: pytest.Session):
     delete_test_instance()
+    run("docker stop pgtest && docker rm pgtest", shell=True, stdout=DEVNULL)  # noqa: S602
+    ln.setup.settings.auto_connect = AUTO_CONNECT
 
 
 @pytest.fixture
